Move firebaseConfig to common module
