+++ app/js/jointeam.js
@@ -1,16 +1,5 @@
 const { dialog } = require('electron').remote;
-
-/** Firebase Config */
-const firebaseConfig = {
-  apiKey: 'AIzaSyBmn_tDSlm4lLdrvSqj8Yb00KkYae8cL-Y',
-  authDomain: 'neon-pulse-development.firebaseapp.com',
-  databaseURL: 'https://neon-pulse-development.firebaseio.com',
-  projectId: 'neon-pulse-development',
-  storageBucket: 'neon-pulse-development.appspot.com',
-  messagingSenderId: '240091062123',
-  appId: '1:240091062123:web:babe11f5f03ced38fbb62e',
-  measurementId: 'G-VMS6JL8H4S',
-};
+const { firebaseConfig } = require('./js/common.js');
 
 /** Initialize Firebase */
 firebase.initializeApp(firebaseConfig);
