+++ app/js/signup.js
@@ -1,15 +1,6 @@
 const { dialog } = require('electron').remote;
+const { firebaseConfig } = require('./js/common.js');
 
-const firebaseConfig = {
-  apiKey: 'AIzaSyBmn_tDSlm4lLdrvSqj8Yb00KkYae8cL-Y',
-  authDomain: 'neon-pulse-development.firebaseapp.com',
-  databaseURL: 'https://neon-pulse-development.firebaseio.com',
-  projectId: 'neon-pulse-development',
-  storageBucket: 'neon-pulse-development.appspot.com',
-  messagingSenderId: '240091062123',
-  appId: '1:240091062123:web:babe11f5f03ced38fbb62e',
-  measurementId: 'G-VMS6JL8H4S',
-};
 // Initialize Firebase
 firebase.initializeApp(firebaseConfig);
