+++ app/signin.html
@@ -19,7 +19,7 @@
         <li>By proceeding you also agree to the Terms of Service</li>
         <li>and Privacy Policy</li>
     </ul>
-    <script src="js/js_functions/signin_functions.js"></script>
+    <script src="./js/js_functions/signin_functions.js"></script>
     <script src="./js/signin.js"></script>
 </body>
